chore(ui): consolidate markdown files
