+++ ui/src/assets/docs/basic.md
