+++ ui/src/assets/docs/dashboard_home.md
